Create RSA_Stepwise_Excl
This script reads in data generated with RSA_sparse_sim and examines the effect of stepwise exclusion of neurons. You will see that based on whether the index is generated by a sparse set of neurons or whether all neurons contributed to the index different functions arise (a step function in the former case, and a linear function in the latter case).